Explicit cast to qint64 in order to make a 64-bit build work correctly

+++ src/webappmanagerservice.cpp
@@ -281,7 +281,7 @@ bool WebAppManagerService::listRunningApps(LSMessage &message)
     Q_FOREACH(WebApplication *app, mWebAppManager->applications()) {
         QJsonObject appObj;
         appObj.insert("appId", app->id());
-        appObj.insert("processId", app->processId());
+        appObj.insert("processId", (qint64) app->processId());
         runningApps.append(QJsonValue(appObj));
     }
